[BUGFIX][BREAKING] Stop adding the `decorators` Babel plugin
In the most recent versions of `ember-cli-babel` we started including
the legacy decorators transform by default. When used with this plugin,
this causes breakage, because both plugins are adding _conflicting_
internal babel plugins.

In general, this is problematic, because the `filter-imports` plugin is
making assumptions about which version of decorators will be used. It
would instead make more sense to allow the user to choose which version
they want by adding and configuring the
`@babel/plugin-proposal-decorators` package themselves.

This PR removes the plugin, and adds the proposal package for local
testing. Users of the package _should_ already be required to add the
decorators transform if they are filtering decorators, so this shouldn't
be an issue in general, but this is _technically_ breaking so I think it
would make sense to bump a major version.